refactor: rename parts of day to schedule
