+++ src/app/components/schedule/schedule.component.spec.ts
@@ -2,16 +2,15 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ScheduleComponent } from './schedule.component';
 
-describe('ScheduleComponent', () => {
+describe('PartsOfDayComponent', () => {
   let component: ScheduleComponent;
   let fixture: ComponentFixture<ScheduleComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [ScheduleComponent]
-    })
-    .compileComponents();
-    
+      imports: [ScheduleComponent],
+    }).compileComponents();
+
     fixture = TestBed.createComponent(ScheduleComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
