+++ src/app/pages/prescription/stp4-add-medication/stp4-add-medication.component.html
@@ -31,12 +31,12 @@
       <!-- <app-medication-search></app-medication-search> -->
 
       <!-- part of day -->
-      <app-parts-of-day
+      <app-schedule
         (filteredPartsOfDayHoursChange)="HandleScheduleChange($event)"
         [partsOfDayHours]="selectedMedication.administrationHours"
         [isDispenseQuantityReadOnly]="false"
         [showEmptyCases]="true"
-      ></app-parts-of-day>
+      ></app-schedule>
       <!-- period -->
       <div class="border border-black p-2 mt-1 d-flex gap-1 gap-lg-3 justify-content-evenly position-relative align-items-center">
         <p class="position-absolute px-2 bg-white fit-content" style="z-index: 2; top: -0.8em; left: 1em">Period*</p>
